Changed the columns
As proposed in #9, made the settings page a little more readable.  It's
can still be made better, but this is a start. Also, removed the fixed
window dimentsions on line (now) 343. Why was that there in the first
place? AHK can figure out the dimensions itself, so this will be a more
flexible approach.